+++ nextgen/config/src/workspace/codeowners_config.rs
@@ -1,0 +1,23 @@
+use moon_common::cacheable;
+use rustc_hash::FxHashMap;
+use schematic::{derive_enum, Config, ConfigEnum};
+
+derive_enum!(
+    #[derive(ConfigEnum, Copy, Default)]
+    pub enum CodeownersOrderBy {
+        #[default]
+        FileSource,
+        ProjectName,
+    }
+);
+
+cacheable!(
+    #[derive(Config)]
+    pub struct CodeownersConfig {
+        pub global_paths: FxHashMap<String, Vec<String>>,
+
+        pub order_by: CodeownersOrderBy,
+
+        pub sync_on_run: bool,
+    }
+);