+++ nextgen/config/src/workspace/vcs_config.rs
@@ -9,12 +9,30 @@ derive_enum!(
     }
 );
 
+derive_enum!(
+    #[derive(ConfigEnum, Copy, Default)]
+    pub enum VcsProvider {
+        Bitbucket,
+
+        #[default]
+        #[serde(rename = "github")]
+        GitHub,
+
+        #[serde(rename = "gitlab")]
+        GitLab,
+
+        Other,
+    }
+);
+
 #[derive(Clone, Config)]
 pub struct VcsConfig {
     #[setting(default = "master")]
     pub default_branch: String,
 
     pub manager: VcsManager,
+
+    pub provider: VcsProvider,
 
     #[setting(default = vec!["origin".into(), "upstream".into()])]
     pub remote_candidates: Vec<String>,